Start process to give ownership of profiles/rounds
Start the process of giving ownership of profiles and rounds to the user
who creates specific profiles and rounds. Change controllers of profiles
and rounds to inherit from ProtectedController. Change set_profile and
set_round to include current user and in profile and round params, make
:profile and :round required.

Create curl script for testing POST for to /profiles.